Split CloneCommand and ChallengeCloneCommand
